check only multiple events
